+++ tests/mocks_test.py
@@ -2,7 +2,6 @@ from urllib3_mock import Responses
 import demisto_client
 from datetime import datetime
 import json
-
 responses = Responses('requests.packages.urllib3')
 
 api_key = 'sample_api_key'
@@ -16,6 +15,7 @@ def assert_reset():
 
 def test_get_docker_images():
     '''Check GET docker images.'''
+
     @responses.activate
     def run():
         body = '{ "images": [{"id": "aae7b8aaba8c", "repository": ' \
@@ -45,6 +45,7 @@ def test_get_docker_images():
 
 def test_create_incident():
     '''Check creating an incident.'''
+
     @responses.activate
     def run():
         body = '{"name":"Test Incident","owner":"Admin","parent":"","phase":"",' \
@@ -88,6 +89,7 @@ def test_create_incident():
 
 def test_get_reports():
     '''Testing GET all reports.'''
+
     @responses.activate
     def run():
         body = '[{"created_by":"DBot","dashboard":"None","decoder":{},"description":"This report ' \
@@ -123,6 +125,7 @@ def test_get_reports():
 
 def test_indicators_search():
     '''Testing search for indicator.'''
+
     @responses.activate
     def run():
         body = r'''
@@ -179,6 +182,7 @@ def test_indicators_search():
 
 def test_export_entry():
     '''Testing export entry artifact.'''
+
     @responses.activate
     def run():
         body = "entry_artifact_6@1770.md"
@@ -207,6 +211,7 @@ def test_export_entry():
 
 def test_generic_request():
     '''Testing generic requst.'''
+
     @responses.activate
     def run():
         responses.add('POST', '/test',
@@ -214,7 +219,8 @@ def test_generic_request():
                       status=200,
                       content_type='text/plain')
         api_instance = demisto_client.configure(base_url=host, api_key=api_key, debug=False)
-        (res, code, headers) = api_instance.generic_request('/test', 'POST', body="this is a test", content_type='text/plain', accept='text/plain')
+        (res, code, headers) = api_instance.generic_request('/test', 'POST', body="this is a test",
+                                                            content_type='text/plain', accept='text/plain')
 
         assert res == 'all good'
         assert code == 200
@@ -226,3 +232,68 @@ def test_generic_request():
 
     run()
     assert_reset()
+
+
+def test_import_layout(mocker):
+    """
+    Given:
+        - Path for a layoutscontainer.
+    When
+        - Using client.import_layout() to upload a new layout.
+    Then
+        - The layout is being uploaded and getting back the layout metadata.
+    """
+    client = demisto_client.configure(base_url=host, api_key=api_key, debug=False,
+                                      verify_ssl=False)
+    mocker.patch.object(client, 'import_layout_with_http_info', return_value={'test': 'test'})
+    res = client.import_layout('tests_data/layoutscontainer-test.json')
+    assert res.get('test') == 'test'
+
+
+def test_import_layout_with_http_info_with_knowing_server_version(mocker):
+    """
+    Given:
+        - path for a layoutscontainer.
+    When
+        - Succeeding in getting demisto version using 'get_layouts_url_for_demisto_version' function
+    Then
+        - The layout is being uploaded and getting back the layout metadata.
+    """
+    client = demisto_client.configure(base_url=host, api_key=api_key, debug=False,
+                                      verify_ssl=False)
+    mocker.patch.object(client.api_client, 'call_api', side_effect=[("{'demistoVersion': '6.0.0'}", 200, {'Content-type': 'application/json'}),  {'test': 'test'}])
+    res = client.import_layout('tests_data/layoutscontainer-test.json')
+    assert res.get('test') == 'test'
+
+
+def test_import_layout_with_http_info_without_knowing_server_version(mocker):
+    """
+    Given:
+        - Path for a layoutscontainer.
+    When
+        - Not knowing demisto version, and working with 6.0.0 or higher
+    Then
+        - The layout is being uploaded and getting back the layout metadata.
+    """
+    client = demisto_client.configure(base_url=host, api_key=api_key, debug=False,
+                                      verify_ssl=False)
+    mocker.patch.object(client.api_client, 'call_api', side_effect=[("{'demistoVersion': '6.0.0'}", 404, {'Content-type': 'application/json'}),  {'test': 'test'}])
+    res = client.import_layout('tests_data/layoutscontainer-test.json')
+    assert res.get('test') == 'test'
+
+
+def test_import_layout_with_http_info_with_old_layout_format(mocker):
+    """
+    Given:
+        - Path for a layout
+    When
+        - Working with 5.0.0 and uploading a layout(the old version)
+    Then
+        - The layout is being uploaded and getting back the layout metadata.
+    """
+    client = demisto_client.configure(base_url=host, api_key=api_key, debug=False,
+                                      verify_ssl=False)
+
+    mocker.patch.object(client.api_client, 'call_api', side_effect=[("{'demistoVersion': '5.0.0'}", 200, {'Content-type': 'application/json'}),  {'test': 'test'}])
+    res = client.import_layout('tests_data/layout-details-test-V2.json')
+    assert res.get('test') == 'test'